separate classes into files, reformat imports
