+++ src/plugins/metal/src/metal_loader.h
@@ -1,6 +1,6 @@
 #pragma once
-#include "halley/concurrency/executor.h"
-#include "halley/core/api/system_api.h"
+#include <halley/concurrency/executor.h>
+#include <halley/core/api/system_api.h>
 
 namespace Halley {
   class MetalLoader
