+++ src/plugins/metal/src/metal_render_target.h
@@ -1,5 +1,5 @@
 #pragma once
-#include "halley/core/graphics/render_target/render_target_screen.h"
+#include <halley/core/graphics/render_target/render_target_screen.h>
 
 namespace Halley {
