+++ src/plugins/metal/src/metal_texture.h
@@ -1,0 +1,23 @@
+#pragma once
+
+#include <halley/core/graphics/texture.h>
+#include <halley/core/graphics/texture_descriptor.h>
+
+#include <Metal/Metal.h>
+
+namespace Halley {
+  class MetalVideo;
+
+  class MetalTexture : public Texture
+  {
+  public:
+    explicit MetalTexture(MetalVideo& video, Vector2i size);
+    void load(TextureDescriptor&& descriptor) override;
+    void bind(id<MTLRenderCommandEncoder> encoder, int bindIndex) const;
+
+  private:
+    MetalVideo& video;
+    id<MTLTexture> metalTexture;
+    id<MTLSamplerState> sampler;
+  };
+}
