+++ src/plugins/metal/src/metal_video.h
@@ -1,24 +1,13 @@
 #pragma once
-#include "metal_shader.h"
 #include "metal_loader.h"
-#include "halley/core/api/halley_api_internal.h"
-#include "halley/core/graphics/texture.h"
-#include "halley/core/graphics/render_target/render_target_texture.h"
-#include "halley/core/graphics/render_target/render_target_screen.h"
-#include "halley/core/graphics/painter.h"
-#include <Metal/Metal.h>
+#include "metal_shader.h"
+#include <halley/core/api/halley_api_internal.h>
+#include <halley/core/graphics/painter.h>
+#include <halley/core/graphics/render_target/render_target_texture.h>
+#include <halley/core/graphics/render_target/render_target_screen.h>
 #include <QuartzCore/CAMetalLayer.h>
 
 namespace Halley {
-
-  static inline MTLRenderPassDescriptor* renderPassDescriptorForTextureAndColour(id<MTLTexture> texture, Colour& colour) {
-    MTLRenderPassDescriptor *pass = [MTLRenderPassDescriptor renderPassDescriptor];
-    pass.colorAttachments[0].clearColor = MTLClearColorMake(colour.r, colour.g, colour.b, colour.a);
-    pass.colorAttachments[0].loadAction  = MTLLoadActionClear;
-    pass.colorAttachments[0].storeAction = MTLStoreActionStore;
-    pass.colorAttachments[0].texture = texture;
-    return pass;
-  }
 
   class MetalVideo final : public VideoAPIInternal
   {
@@ -56,42 +45,4 @@ namespace Halley {
     void initSwapChain(Window& window);
   };
 
-
-  class MetalTexture : public Texture
-  {
-  public:
-    explicit MetalTexture(MetalVideo& video, Vector2i size);
-    void load(TextureDescriptor&& descriptor) override;
-    void bind(id<MTLRenderCommandEncoder> encoder, int bindIndex) const;
-
-  private:
-    MetalVideo& video;
-    id<MTLTexture> metalTexture;
-    id<MTLSamplerState> sampler;
-  };
-
-  class MetalPainter : public Painter
-  {
-  public:
-    explicit MetalPainter(MetalVideo& video, Resources& resources);
-    void clear(Colour colour) override;
-    void setMaterialPass(const Material& material, int pass) override;
-    void doStartRender() override;
-    void doEndRender() override;
-    void setVertices(const MaterialDefinition& material, size_t numVertices, void* vertexData, size_t numIndices, unsigned short* indices, bool standardQuadsOnly) override;
-    void drawTriangles(size_t numIndices) override;
-    void setViewPort(Rect4i rect) override;
-    void setClip(Rect4i clip, bool enable) override;
-    void setMaterialData(const Material& material) override;
-    void onUpdateProjection(Material& material) override;
-
-  private:
-    void setBlending(BlendType blendType, MTLRenderPipelineColorAttachmentDescriptor* colorAttachment);
-    void setBlendFactor(MTLRenderPipelineColorAttachmentDescriptor* colorAttachment, MTLBlendFactor src, MTLBlendFactor dst);
-
-    MetalVideo& video;
-    id<MTLCommandBuffer> buffer;
-    id<MTLRenderCommandEncoder> encoder;
-    id<MTLBuffer> indexBuffer;
-  };
 }
